custom_feature/real_requirements.txt
absl-py==1.0.0
aiohttp==3.8.1
aiosignal==1.2.0
altair==4.1.0
aniso8601==8.0.0
anyio==3.4.0
argcomplete==1.12.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
astor==0.8.1
astroid==2.9.0
async-timeout==4.0.1
attrs==21.2.0
autopep8==1.5.5
Babel==2.9.1
backcall==0.2.0
backports.zoneinfo==0.2.1
base58==2.1.1
bleach==4.1.0
blinker==1.4
bokeh==2.4.2
cachetools==4.2.4
catfish==4.16.3
certifi==2021.10.8
cffi==1.15.0
chardet==3.0.4
charset-normalizer==2.0.9
cheetah @ file:///tmp/cheetah-0.4.16-py3-none-any.whl
click==7.1.2
cloudpickle==2.0.0
convcolors==2.2.0
cycler==0.11.0
Cython==0.29.25
dbus-python==1.2.16
debugpy==1.5.1
decorator==5.1.0
defusedxml==0.7.1
entrypoints==0.3
extcolors==1.0.0
faiss-gpu==1.7.2
flake8==3.8.4
Flask==1.1.1
Flask-HTTPAuth==3.3.0
Flask-RESTful==0.3.8
Flask-Script==2.0.6
Flask-Testing==0.7.1
fonttools==4.28.3
frozenlist==1.2.0
fsspec==2021.11.1
future==0.18.2
gitdb==4.0.9
GitPython==3.1.24
google-auth==2.3.3
google-auth-oauthlib==0.4.6
grpcio==1.42.0
h5py==3.6.0
idna==3.3
imageio==2.13.3
importlib-metadata==4.8.2
importlib-resources==5.4.0
imutils==0.5.4
iniconfig==1.1.1
ipykernel==6.6.0
ipython==7.30.1
ipython-genutils==0.2.0
ipywidgets==7.6.5
isort==5.10.1
itsdangerous==1.1.0
jedi==0.17.2
Jinja2==3.0.3
JLDracula==0.1.0
joblib==1.1.0
json5==0.9.6
JsonForm==0.0.2
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.2.1
JsonSir==0.0.2
jupyter-bokeh==3.0.4
jupyter-client==7.1.0
jupyter-core==4.9.1
jupyter-lsp==1.5.1
jupyter-resource-usage==0.6.1
jupyter-server==1.13.1
jupyter-server-proxy==3.2.0
jupyterlab==3.1.7
jupyterlab-execute-time==2.1.0
jupyterlab-fasta==3.1.1
jupyterlab-geojson==3.1.2
jupyterlab-horizon-theme==3.0.1
jupyterlab-katex==3.2.2
jupyterlab-lsp==3.9.2
jupyterlab-mathjax3==4.2.2
jupyterlab-nvdashboard==0.6.0
jupyterlab-pygments==0.1.2
jupyterlab-server==2.8.2
jupyterlab-system-monitor==0.8.0
jupyterlab-topbar==0.6.1
jupyterlab-vega2==3.1.3
jupyterlab-vega3==3.1.3
jupyterlab-widgets==1.0.2
kiwisolver==1.3.2
lazy-object-proxy==1.6.0
lckr-jupyterlab-variableinspector==3.0.9
Markdown==3.3.6
MarkupSafe==2.0.1
matplotlib==3.5.1
matplotlib-inline==0.1.3
mccabe==0.6.1
mistune==0.8.4
mpmath==1.2.1
multidict==5.2.0
nbclassic==0.3.4
nbclient==0.5.9
nbconvert==6.3.0
nbformat==5.1.3
nest-asyncio==1.5.4
networkx==2.6.3
notebook==6.4.6
numpy==1.21.4
oauthlib==3.1.1
opencv-contrib-python==4.5.4.60
packaging==21.3
pandas==1.3.5
pandocfilters==1.5.0
parso==0.7.1
patsy==0.5.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.4.0
pipx==0.16.4
pkg-resources==0.0.0
platformdirs==2.4.0
plotly==5.4.0
pluggy==1.0.0
prometheus-client==0.12.0
prompt-toolkit==3.0.24
protobuf==3.19.1
psutil==5.8.0
ptyprocess==0.7.0
py==1.11.0
pyarrow==6.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.16.2
pycodestyle==2.6.0
pycparser==2.21
pycryptodome==3.12.0
pydeck==0.7.1
pyDeprecate==0.3.1
pydocstyle==6.1.1
pyflakes==2.2.0
Pygments==2.10.0
PyGObject==3.36.0
PyJWT==1.7.1
pylint==2.12.2
Pympler==0.9
pynvml==11.4.1
pyparsing==3.0.6
pyrsistent==0.18.0
pytest==6.2.5
python-apt==2.0.0+ubuntu0.20.4.6
python-dateutil==2.8.2
python-debian===0.1.36ubuntu1
Python-EasyConfig==0.1.7
python-jsonrpc-server==0.4.0
python-language-server==0.36.2
pytorch-lightning==1.5.5
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
PyWavelets==1.2.0
PyYAML==6.0
pyzmq==22.3.0
requests==2.26.0
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
requests-unixsocket==0.2.0
Resource==0.2.1
rope==0.22.0
rsa==4.8
scikit-image==0.19.0
scikit-learn==1.0.1
scipy==1.7.3
seaborn==0.11.2
Send2Trash==1.8.0
simpervisor==0.4
six==1.16.0
smmap==5.0.0
sniffio==1.2.0
snowballstemmer==2.2.0
statsmodels==0.13.1
streamlit==1.2.0
sympy==1.9
tenacity==8.0.1
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorrt==8.0.1.6
terminado==0.12.1
testpath==0.5.0
threadpoolctl==3.0.0
tifffile==2021.11.2
toml==0.10.2
toolz==0.11.2
torch==1.10.0+cu113
torchaudio==0.10.0+cu113
torchmetrics==0.6.1
torchnet==0.0.4
torchvision==0.11.1+cu113
tornado==6.1
tqdm==4.62.3
traitlets==5.1.1
typing-extensions==4.0.1
tzdata==2021.5
tzlocal==4.1
ujson==4.3.0
urllib3==1.26.7
userpath==1.7.0
validators==0.18.2
visdom==0.2.4
watchdog==2.1.6
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.2.3
Werkzeug==1.0.0
widgetsnbextension==3.5.2
wrapt==1.13.3
yapf==0.31.0
yarl==1.7.2
zipp==3.6.0
